requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
python-socketio==5.8.0
eventlet==0.33.3
Pillow==10.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
Werkzeug==2.3.7
redis==5.0.1
celery==5.3.4
requests==2.31.0
email_validator==2.2.0 